Create TripAdvisor_Scrapping
Web Scrapping of Trip Advisor Website is performed at 2 levels of depth horizontally and 2 levels depth at vertical level. 
The goal is to extract statistics for all holiday destinations in the first 40 homepages. 
Extracting names of the destination, the rating that travelers have given, the review or comment left by each person who made use of the service, the name of the author, etc. 
To later perform a data analysis using PowerBI